doslib: move call_msdos() to msdoshlp.c
