+++ src/dosext/dpmi/msdoshlp.h
@@ -72,6 +72,7 @@ struct pmaddr_s doshlp_get_entry32(unsigned entry);
 
 void doshlp_quit_dpmi(cpuctx_t *scp);
 struct pmaddr_s doshlp_get_abort_helper(void);
+void doshlp_call_msdos(cpuctx_t *scp);
 void doshlp_call_reinit(cpuctx_t *scp);
 int doshlp_idle(void);
